conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.1
anaconda_prompt=1.1.0=haa95532_1
asgiref=3.8.1=py313haa95532_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.12.2=haa95532_0
django=5.2=py313haa95532_0
djangorestframework=3.16.1=pyhd8ed1ab_0
expat=2.7.1=h8ddb27b_0
libffi=3.4.4=hd77b12b_1
libmpdec=4.0.0=h827c3e9_0
libzlib=1.3.1=h02ab6af_0
openssl=3.0.18=h543e019_0
pip=25.2=pyhc872135_0
python=3.13.7=h260b955_100_cp313
python_abi=3.13=1_cp313
setuptools=78.1.1=py313haa95532_0
sqlite=3.50.2=hda9a48d_1
sqlparse=0.5.2=py313haa95532_0
tk=8.6.15=hf199647_0
typing_extensions=4.15.0=py313haa95532_0
tzdata=2025b=h04d1e81_0
ucrt=10.0.22621.0=haa95532_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
waitress=3.0.2=py313haa95532_0
wheel=0.45.1=py313haa95532_0
xz=5.6.4=h4754444_1
zlib=1.3.1=h02ab6af_0
